# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Naam Jaap Sewa Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jaap/count:
    post:
      tags: [Jaap]
      operationId: addJaapCount
      summary: Increment today's jaap count
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddJaapBody'
      responses:
        '200':
          description: Updated jaap snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JaapSnapshotEnvelope'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Sankalp not accepted today.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limited / suspicious activity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /jaap/snapshot:
    get:
      tags: [Jaap]
      operationId: getJaapSnapshot
      summary: Get the current user's jaap snapshot
      responses:
        '200':
          description: Snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JaapSnapshotEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /payouts:
    get:
      tags: [Profile]
      operationId: listMyPayouts
      summary: Current user's payout history
      responses:
        '200':
          description: Payouts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutsEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Profile]
      operationId: requestPayout
      summary: Request a payout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayoutRequest'
      responses:
        '200':
          description: Payout requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutEnvelope'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PatronSankalpProgress:
      type: object
      required:
        - id
        - goalCount
        - accumulated
        - userAccumulated
        - remaining
        - yajamanaName
        - purpose
        - ratePerJaap
      properties:
        id:
          type: string
        goalCount:
          type: integer
        accumulated:
          type: integer
        userAccumulated:
          type: integer
        remaining:
          type: integer
        yajamanaName:
          type: string
        purpose:
          type: string
        ratePerJaap:
          type: number
    JaapSnapshot:
      type: object
      required:
        - todayCount
        - totalCount
        - todayEarnings
        - totalEarnings
        - sankalpAccepted
        - streakDays
        - todayMalas
        - totalMalas
        - allSankalpsDone
      properties:
        todayCount:
          type: integer
        totalCount:
          type: integer
        todayEarnings:
          type: number
        totalEarnings:
          type: number
        sankalpAccepted:
          type: boolean
        streakDays:
          type: integer
        todayMalas:
          type: integer
        totalMalas:
          type: integer
        activePatronSankalpId:
          type: string
          nullable: true
        currentSankalp:
          oneOf:
            - $ref: '#/components/schemas/PatronSankalpProgress'
            - type: 'null'
        activeMantraText:
          type: string
          nullable: true
        allSankalpsDone:
          type: boolean
        pendingSankalpAcceptance:
          type: object
          nullable: true
    JaapSnapshotEnvelope:
      type: object
      required: [snapshot]
      properties:
        snapshot:
          $ref: '#/components/schemas/JaapSnapshot'
    AddJaapBody:
      type: object
      required: [count, intervalMs]
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 200
        intervalMs:
          type: integer
          minimum: 0
        verified:
          type: boolean
...
    Payout:
      type: object
      required: [id, userId, amount, status, requestedAt, devoteeName]
      properties:
        id:
          type: string
        userId:
          type: string
        devoteeName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, rejected]
        upiId:
          type: [string, 'null']
        paymentMethod:
          type: [string, 'null']
        paymentNote:
          type: [string, 'null']
        requestedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, 'null']
          format: date-time
    PayoutEnvelope:
      type: object
      required: [payout]
      properties:
        payout:
          $ref: '#/components/schemas/Payout'
    PayoutsEnvelope:
      type: object
      required: [payouts]
      properties:
        payouts:
          type: array
          items:
            $ref: '#/components/schemas/Payout'
    PayoutRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 10
        upiId:
          type: string
          minLength: 3
...